Update .cpanel.yml
Previous attempt to add 'html' director did not work - I am trying again but relocating the code above the commands to add individual files to ensure order of operations is not the issue.  I have also added a 'css' directory command to give me more to play with - I have not yet created the css directory in my Git repo so am interested to see whether anything results from this.